+++ packages/devtools_app/lib/src/shared/config_specific/launch_url/launch_url.dart
@@ -5,9 +5,8 @@
 import 'package:url_launcher/url_launcher.dart' as url_launcher;
 
 import '../../globals.dart';
-import '_launch_url_stub.dart'
-    if (dart.library.js_interop) '_launch_url_web.dart'
-    if (dart.library.io) '_launch_url_desktop.dart';
+import '_launch_url_desktop.dart'
+    if (dart.library.js_interop) '_launch_url_web.dart';
 
 Future<void> launchUrl(String url) async {
   final parsedUrl = Uri.tryParse(url);