+++ packages/devtools_app/lib/src/shared/config_specific/logger/allowed_error.dart
@@ -2,5 +2,5 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-export 'allowed_error_default.dart'
-    if (dart.library.js_interop) 'allowed_error_html.dart';
+export '_allowed_error_default.dart'
+    if (dart.library.js_interop) '_allowed_error_web.dart';